+++ dashboard/lib/hooks/use-domain.ts
@@ -4,14 +4,28 @@ import { querySQL } from '../api'
 import { DomainData, DomainQueryData } from '../types/domain'
 
 async function getDomain(): Promise<DomainData> {
-  const { data } = await querySQL<DomainQueryData>(
-    'SELECT href FROM analytics_hits ORDER BY timestamp DESC LIMIT 1 FORMAT JSON'
-  )
-  const results =
-    /^(?:https?:\/\/)?(?:[^@\n]+@)?(?:www\.)?([^:\/\n\?\=]+)/g.exec(
-      data[0]['href']
-    )
-  const domain = results?.length ? results[1] : null
+  // Guess the instrumented domain, and exclude other domains like development or staging.
+  //  - Try to get the domain with most hits from the last hour.
+  //  - Fallback to 'some' domain.
+  // Best balance between data accuracy and performance I can get.
+  const { data } = await querySQL<DomainQueryData>(`
+    with (
+      SELECT nullif(domainWithoutWWW(href),'') as domain
+      FROM analytics_hits
+      where timestamp >= now() - interval 1 hour
+      group by domain
+      order by count(1) desc
+      limit 1
+    ) as top_domain,
+    (
+      SELECT domainWithoutWWW(href)
+      FROM analytics_hits
+      where href not like '%localhost%'
+      limit 1
+    ) as some_domain
+    select coalesce(top_domain, some_domain) as domain format JSON
+  `)
+  const domain = data[0]['domain'];
   const logo = domain
     ? `https://s2.googleusercontent.com/s2/favicons?domain=${domain}`
     : FALLBACK_LOGO