+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
 - [Script] Remove source maps
+- [Dashboard] Improve domain guessing
 
 ## [1.2.0] - 2022-09-06
 ### Changed